add more status tips

+++ src/himena_relion/_utils.py
@@ -428,28 +428,33 @@ def open_with_lock(
     lock_dir = pipeline_path.parent / ".relion_lock"
     each_wait = 0.05
     num_trial = int(wait_sec / each_wait) + 1
-    for _ in range(num_trial):
-        try:
-            lock_dir.mkdir(exist_ok=False)
-        except FileNotFoundError:
-            raise  # FileNotFoundError is a subclass of OSError.
-        except OSError:
-            time.sleep(each_wait)
-        else:
-            break
-    else:
-        raise RelionPipelineLockError(
-            f"Failed to acquire lock for {pipeline_path}. Another instance may be "
-            "editing the pipeline, or the previous run may have crashed. "
-        )
     try:
-        with pipeline_path.open(mode) as f:
-            yield f
+        for _ in range(num_trial):
+            try:
+                lock_dir.mkdir(exist_ok=False)
+            except FileNotFoundError:
+                raise  # FileNotFoundError is a subclass of OSError.
+            except OSError:
+                time.sleep(each_wait)
+            else:
+                break
+        else:
+            raise RelionPipelineLockError(
+                f"Failed to acquire lock for {pipeline_path}. Another instance may be "
+                "editing the pipeline, or the previous run may have crashed. "
+            )
+        try:
+            with pipeline_path.open(mode) as f:
+                yield f
+        finally:
+            # remove the lock
+            with suppress(Exception):
+                lock_dir.rmdir()
+            pipeline_path.touch()
     finally:
-        # remove the lock
-        with suppress(Exception):
-            lock_dir.rmdir()
-        pipeline_path.touch()
+        if lock_dir.exists():
+            with suppress(Exception):
+                lock_dir.rmdir()
 
 
 def extract_input_edges(params: dict[str, str], keys: Iterable[str]) -> list[str]:
